ClimateModels.jl

This package provides a uniform interface to climate models of varying complexity and completeness. Models that range from low dimensional to whole Earth System models are ran and analyzed via a simple interface laid out hereafter. Three examples illustrate this framework as applied to:

one stochastic path (0D)

a shallow water model (2D)

the MIT general circulation model (3D, Ocean, Atmosphere, etc)

Note: this package is still its early stage of development, such that breaking changes should be expected.

ClimateModels.ModelConfig

ClimateModels.add_git_msg

ClimateModels.build

ClimateModels.clean

ClimateModels.compile

ClimateModels.init_git_log

ClimateModels.launch

ClimateModels.monitor

ClimateModels.setup

Climate Model Interface

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: Array{String,1} = Array{String,1}(undef, 0)
inputs :: Array{String,1} = Array{String,1}(undef, 0)
outputs :: Array{String,1} = Array{String,1}(undef, 0)
status :: Array{String,1} = Array{String,1}(undef, 0)
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and clean the run directory (via rm)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0,configuration="anonymous",options=Array{String,1}(undef, 0))
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.compile — Function

compile(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0)
setup(tmp)
compile(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
launch(tmp)

ClimateModels.monitor — Function

monitor(x)

Show x.status[end] by default.

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
monitor(tmp)

Git Support

ClimateModels.init_git_log — Function

init_git_log(x :: AbstractModelConfig)

Create log subfolder, initialize git, and commit initial README.md

ClimateModels.add_git_msg — Function

add_git_msg(x :: AbstractModelConfig,msg,commit_msg)

Add message msg to the log/README.md file and git commit.