API Reference

Data Structure

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

Generic data structure for a model configuration. This serves as :

default concrete type for AbstractModelConfig

keyword constructor for AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.ModelConfig — Method

ModelConfig(func::Function,inputs::NamedTuple)

Simplified constructor for case when model is a Function.

Methods

ClimateModels.setup — Method

setup(x::AbstractModelConfig)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig) # hide

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)
build(tmp)
launch(tmp)

Base.log — Function

log(x :: AbstractModelConfig)

Show the record of git commits that have taken place in the log folder.

log(x :: AbstractModelConfig, commit_msg :: String; 
             fil="", msg="", init=false, prm=false)

Keyword arguments work like this

init==true : create log subfolder, initialize git, and commit initial README.md

prm==true : add files found in input or tracked_parameters/ (if any) to git log

!isempty(fil) : commit changes to file log/$(fil) with message commit_msg. If log/$(fil) is unknown to git (i.e. commit errors out) then try adding log/$(fil) first.

and are mutually exclusive (i.e., use only one at a time).

MC=run(ModelConfig(ClimateModels.RandomWalker,(NS=100,)))
MC.inputs[:NS]=200
msg="update tracked_parameters.toml (or skip if up to date)"
log(MC,msg,fil="tracked_parameters.toml",prm=true)
log(MC)

Simplified API

ClimateModels.@ModelRun — Macro

@ModelRun(func)

Macro equivalent for run(ModelConfig(model=func)).

@ModelRun(func::AbstractModelConfig)

Macro equivalent for run(ModelConfig(model=func)).

ClimateModels.ModelRun — Function

ModelRun(x :: AbstractModelConfig)

Shorthand for x |> setup |> build |> launch

Returns AbstractModelConfig as output.

Utilities

Base.pathof — Function

pathof(x::AbstractModelConfig)

Returns the run directory path for x ; i.e. joinpath(x.folder,string(x.ID))

Base.Filesystem.readdir — Function

readdir(x::AbstractModelConfig)

Same as readdir(pathof(x)).

readdir(x::AbstractModelConfig,subfolder::String)

Same as readdir(joinpath(pathof(x),subfolder)).

Base.show — Function

show(io::IO, z::AbstractModelConfig)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
show(tmp)

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and rm the run directory (pathof(x))

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

Notebooks

Here are convenience functions to use Pluto.jl notebooks.

ClimateModels.setup — Method

setup(MC::PlutoConfig;IncludeManifest=true)

call default_ClimateModelSetup

call unroll

add notebook_launch to tasks

MC1=PlutoConfig(model="examples/CMIP6.jl")
setup(MC1)
build(MC1)
launch(MC1)

ClimateModels.notebooks.unroll — Function

unroll(PlutoFile::String; EnvPath="", ModuleFile="")

Split up Pluto notebook file (PlutoFile) into main program (main.jl), Project.toml, Manifest.toml, and CellOrder.txt.

these files are saved to folder p (EnvPath or a temporary folder by default)

unroll returns output path p and main program file name

unroll optionally copies companion file mf to p (if file mf exists)

default mf is PlutoFile[1:end-3]*"_module.jl" unless ModuleFile is specified

the reroll function can be used to reassemble as a Pluto notebook

For example:

using Pkg
p,f=notebooks.unroll("CMIP6.jl")
Pkg.activate(p)
Pkg.update()
n=notebooks.reroll(p,f)

ClimateModels.notebooks.reroll — Function

reroll(p,f; PlutoFile="notebook.jl")

The reroll function can be used to reassemble as a Pluto notebook that was previously unroll'ed.

See unroll documentation for an example.

JuliaClimate/Notebooks

Convenience functions for notebooks documented in the JuliaClimate/Notebooks webpage.

ClimateModels.notebooks.list — Function

notebooks.list()

List downloadable notebooks based on the JuliaClimate/Notebooks webpage.

ClimateModels.notebooks.download — Function

notebooks.download(path,nbs)

Download notebooks/files listed in nbs to path.

If nbs.file[i] is found at nbs.url[i] then download it to path/nbs.folder[i].

If a second file is found at nbs.url[i][1:end-3]*"_module.jl" then we download it too.

using ClimateModels, UUIDs
path=joinpath(tempdir(),string(UUIDs.uuid4()))

nbs=notebooks.list()
notebooks.download(path,nbs)

or

using DataFrames
url0="https://raw.githubusercontent.com/JuliaClimate/IndividualDisplacements.jl/master/examples/worldwide/"

nbs2=DataFrame( "folder" => ["IndividualDisplacements.jl","IndividualDisplacements.jl"], 
                "file" => ["ECCO_FlowFields.jl","OCCA_FlowFields.jl"], 
                "url" => [url0*"ECCO_FlowFields.jl",url0*"OCCA_FlowFields.jl"])
notebooks.download(path,nbs2)

Base.open — Function

open(;notebook_path="",notebook_url="",
      pluto_url="http://localhost:1234/",pluto_options="...")

Open notebook in web-browser via Pluto.

Important note: this assumes that the Pluto server is already running, e.g. from Pluto.run(), at URL pluto_url (by default, "http://localhost:1234/", should work on a laptop or desktop).

Examples:

nbs=notebooks.list()
notebooks.open(notebook_url=nbs.url[1])

notebooks.open(notebook_path="examples/defaults.jl")
pluto_url="https://ade.ops.maap-project.org/serverpmohyfxe-ws-jupyter/server-3100/pluto/"
notebooks.open(notebook_path="examples/defaults.jl",pluto_url=pluto_url)

PkgDevConfig

In the package development mode, model is specified as a PackageSpec.

ClimateModels.PkgDevConfig — Function

PkgDevConfig(url::String,func::Function,inputs::NamedTuple)

Simplified constructor for case when model is a url (PackageSpec).

This leads setup to install the chosen package using Pkg.develop. This can be useful for developing a package or using an unregistered package in the context of ClimateModels.jl.

There are two common cases:

if configuration is left undefined then launch will run the package test suite using Pkg.test as in this example (code link)

if configuration is provided as a Function then launch will call it as illustrated in the ShallowWaters.jl example (code link)

Note

As an exercise, can you turn ShallowWaters.jl example into a normal user mode example?