User Manual

Here we document key functionalities offered in ClimateModels.jl

Climate Model Interface

Tracked Worklow Framework

Cloud + On-Premise File Support

Climate Model Interface

The interface ties the ModelConfig data structure with methods like setup, build, and launch. In return, it provides standard methods to deal with inputs and outputs, as well as capabilities described below.

The run method provides the capability to deploy models in streamlined fashion – with just one code line, or just one click. It executes all three steps at once (setup, build, and launch).

With the simplified ModelConfig constructor, we can then just write:

f=ClimateModels.RandomWalker
run(ModelConfig(f))

  ID            = a0bea808-fb57-4866-b947-c323ddedfafc
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/a0bea808-fb57-4866-b947-c323ddedfafc
  log subfolder = /tmp/a0bea808-fb57-4866-b947-c323ddedfafc/log

The above example uses ClimateModels.RandomWalker as the model (function f). By design of our interface, it is required that f receives a ModelConfig as its sole input argument.

Note

In practice, this requirement is easily satisfied. Input parameters can be specified to ModelConfig via the inputs keyword argument, or via files instead.

Often it is most practical to break things down. Let's start with defining the model:

MC=ModelConfig(model=f)

  ID            = 12e08992-aa6a-4b8a-9465-fd0ed436555c
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/12e08992-aa6a-4b8a-9465-fd0ed436555c
  log subfolder = /tmp/12e08992-aa6a-4b8a-9465-fd0ed436555c/log

The model's top level function gets called via launch. In our example, f thus generates a file called RandomWalker.csv, which gets stored in the run folder.

The run sequence is shown below. In practice, setup typically handles files and software, build may compile a chosen model configuration, and launch takes care of the main computation.

setup(MC)
build(MC)
launch(MC)

Note

Compilation during build is not a requirement. It can also be done within launch or beforehand.

Sometimes it's convenient to further break down the computational workflow into several tasks. These can be added to the ModelConfig via put! and then executed via launch, as demonstrated below.

The run folder name and its content can be viewed using pathof and readdir, respectively.

pathof(MC)

"/tmp/12e08992-aa6a-4b8a-9465-fd0ed436555c"

readdir(MC)

2-element Vector{String}:
 "RandomWalker.csv"
 "log"

The log subfolder was created earlier by setup. The log function retrieves the workflow log.

log(MC)

5-element Vector{String}:
 "797e337 initial setup"
 "d7b17f1 add Project.toml to log"
 "6654aa9 add Manifest.toml to log"
 "9e7054d task started [35fb971c-64c7-44ee-93a0-faa26e0168ba]"
 "2a6f55d (HEAD -> main) task ended   [35fb971c-64c7-44ee-93a0-faa26e0168ba]"

This highlights that Project.toml and Manifest.toml for the environment being used have been archived. This happens during setup to document all dependencies and make the workflow reproducible.

Customization

A key point is that everything can be customized to, e.g., use popular models previously written in Fortran or C just as simply.

The simplest way to use the ClimateModels.jl interface is to specify model directly as a function, and use defaults for everything else, as illustrated in random walk. Alternatively, the model name can be provided as a String and the main Function as the configuration, as in CMIP6.

Often, however, one may want to define custom setup, build, or launch methods. To this end, one can define a concrete type of AbstractModelConfig using ModelConfig as a blueprint. This is the recommended approach when other languanges like Fortran or Python are involved (e.g., Hector, FaIR, SPEEDY, MITgcm).

Note

Defining a concrete type of AbstractModelConfig can also be practical with pure Julia model, e.g. to speed up launch, generate ensembles, facilitate checkpointing, etc. That's the case in the Oceananigans.jl example.

For popular models the customized interface elements can be provided via a dedicated package. This may allow them to be maintained independently by developers and users most familiar with each model. MITgcmTools.jl does this for MITgcm. It provides its own suite of examples that use the ClimateModels.jl interface.

Additional Example

In this example, we illustrate how one can interact with model parameters and rerun models. After an initial model run of 100 steps, duration NS is extended to 200 time steps. The put! and launch sequence then reruns the model.

The same method can be used to break down a workflow in several steps. Each call to launch sequentially takes the next task from the stack (i.e., channel). Once the task channel is empty then launch does nothing.

Note

The call sequence is readily reflected in the workflow log, and the run dir now has two output files. The modified parameters are also automatically recorded in tracked_parameters.toml during launch.

MC=ModelConfig(f,(NS=100,filename="run01.csv"))
run(MC)

MC.inputs[:NS]=200
MC.inputs[:filename]="run02.csv"
put!(MC)
launch(MC)

log(MC)

9-element Vector{String}:
 "6919192 initial setup"
 "2891563 initial tracked_parameters.toml"
 "f978eda add Project.toml to log"
 "eee0a52 add Manifest.toml to log"
 "e813342 task started [8d5d3bd4-b915-4c09-8250-82a580d731ce]"
 "36796e4 task ended   [8d5d3bd4-b915-4c09-8250-82a580d731ce]"
 "960f488 task started [55e34303-2bf9-429a-9b7b-4c4182f05bc5]"
 "c015c72 modify tracked_parameters.toml"
 "e290964 (HEAD -> main) task ended   [55e34303-2bf9-429a-9b7b-4c4182f05bc5]"

readdir(MC)

4-element Vector{String}:
 "log"
 "run01.csv"
 "run02.csv"
 "tmp.txt"

Tracked Worklow Support

The setup method normally calls log to create a temporary run folder with a git enabled subfolder called log. This allows for recording each workflow step, using log functions listed below.

Calling log on a ModelConfig without any other argument shows the workflow record. This feature is as illustrated several times in the above examples.

File and Cloud Support

There are various ways that numerical model output gets archived, distributed, and retrieved from the internet. In some cases downloading data can be the most convenient approach. In others it can be more advantageous to compute in the cloud and only download final results for plotting.

ClimateModels.jl comes equiped with packages that read popular file formats used in climate modeling and science. Downloads.jl, CSV.jl, DataFrames.jl, NetCDF.jl, Zarr.jl, and TOML.jl are thus readily available when you install ClimateModels.jl. For instance, one can read the CSV file generated before as

fil=joinpath(pathof(MC),"RandomWalker.csv")
CSV=ClimateModels.CSV # hide
DataFrame=ClimateModels.DataFrame #hide
CSV.File(fil) |> DataFrame
summary(ans) # hide

For additional examples covering other file formats, please refer to the IPCC report and CMIP6 archive notebooks and code links.

Data Structure

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.ModelConfig — Method

ModelConfig(func::Function,inputs::NamedTuple)

Simplified constructor for case when model is a Function.

ClimateModels.PkgDevConfig — Function

PkgDevConfig(url::String,func::Function,inputs::NamedTuple)

Simplified constructor for case when model is a url (PackageSpec).

Methods

ClimateModels.setup — Function

setup(MC::AbstractModelConfig,PlutoFile::String)

Call default setup then

Call notebooks.unroll

Consolidate main.jl (activate, instantiate)

MC1=ModelConfig()
notebooks.setup(MC1,"examples/CMIP6.jl")

cd(joinpath(pathof(MC1),"run"))
include("main.jl")

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig) # hide

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)
build(tmp)
launch(tmp)

Base.run — Function

run(x :: AbstractModelConfig)

Shorthand for setup |> build |> launch, which returns AbstractModelConfig as output.

Base.log — Function

log(x :: AbstractModelConfig)

Show the record of git commits that have taken place in the log folder.

log(x :: AbstractModelConfig, commit_msg :: String; 
             fil="", msg="", init=false, prm=false)

Keyword arguments work like this

init==true : create log subfolder, initialize git, and commit initial README.md

prm==true : add files found in input or tracked_parameters/ (if any) to git log

!isempty(fil) : commit changes to file log/$(fil) with message commit_msg. If log/$(fil) is unknown to git (i.e. commit errors out) then try adding log/$(fil) first.

and are mutually exclusive (i.e., use only one at a time).

MC=run(ModelConfig(ClimateModels.RandomWalker,(NS=100,)))
MC.inputs[:NS]=200
msg="update tracked_parameters.toml (or skip if up to date)"
log(MC,msg,fil="tracked_parameters.toml",prm=true)
log(MC)

Utilities

Base.pathof — Function

pathof(x::AbstractModelConfig)

Returns the run directory path for x ; i.e. joinpath(x.folder,string(x.ID))

Base.Filesystem.readdir — Function

readdir(x::AbstractModelConfig)

Same as readdir(pathof(x)).

readdir(x::AbstractModelConfig,subfolder::String)

Same as readdir(joinpath(pathof(x),subfolder)).

Base.show — Function

show(io::IO, z::AbstractModelConfig)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
show(tmp)

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and rm the run directory (pathof(x))

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

Notebooks

Missing docstring.

Missing docstring for notebooks.setup. Check Documenter's build log for details.

ClimateModels.notebooks.unroll — Function

unroll(PlutoFile::String; EnvPath="")

Extract main program, Project.toml, and Manifest.toml from Pluto notebook file PlutoFile. Save them in folder EnvPath (default = temporary folder). Typical use case is shown below.

p,f=notebooks.unroll("CMIP6.jl")
cd(p)
Pkg.activate("./")
Pkg.instantiate()
include(f)

Base.open — Function

open(;notebook_path="",notebook_url="",
      pluto_url="http://localhost:1234/",pluto_options="...")

Open notebook in web-browser via Pluto. Important note: this assumes that the Pluto server is already running, e.g. from Pluto.run(), at URL pluto_url (by default, "http://localhost:1234/", should work on a laptop or desktop).

Simple examples:

notebooks.open(notebook_path="examples/defaults.jl")

nbs=notebooks.list()
notebooks.open(notebook_url=nbs.url[1])

More examples:

nbs=notebooks.list()
path=joinpath(tempdir(),"nbs")
notebooks.download(path,nbs)

pluto_url="https://ade.ops.maap-project.org/serverpmohyfxe-ws-jupyter/server-3100/pluto/"

ii=1
notebook_path=joinpath(path,nbs.folder[ii],nbs.file[ii])
notebooks.open(pluto_url=pluto_url,notebook_path=notebook_path)

Package Development Mode

The defining feature of this alternative method is specifying model as a PackageSpec. This makes setup install the chosen package using Pkg.develop. This allows for developing a package or using an unregistered package in the context of ClimateModels.jl.

There are two common cases:

if configuration is left undefined then launch will run the package test suite using Pkg.test as in this example (code link)

if configuration is provided as a Function then launch will call it as illustrated in the ShallowWaters.jl example (code link)

Note

As an exercise, can you turn ShallowWaters.jl example into a normal user mode example?