ClimateModels.jl

This package provides a uniform interface to climate models of varying complexity and completeness. Models that range from low dimensional to whole Earth System models can be run and/or analyzed via this framework.

ClimateModels.jl is also aimed at cloud computing workflows that often start from previous model output available over the internet. Version control support is provided via Git.jl to facilitate documentation and reproducibility.

Examples That Run Models

one stochastic path (0D)

a shallow water model (2D)

the MIT general circulation model (3D, Ocean, Atmosphere, etc)

Remote Access To Model Output

The initial example accesses CMIP6 model output from cloud storage, and computes time-mean maps + global-mean time series of temperature.

Note: this package is still its early stage of development, such that breaking changes should be expected.

ClimateModels.ModelConfig

ClimateModels.build

ClimateModels.clean

ClimateModels.cmip

ClimateModels.compile

ClimateModels.git_log_fil

ClimateModels.git_log_msg

ClimateModels.git_log_prm

ClimateModels.init_git_log

ClimateModels.launch

ClimateModels.monitor

ClimateModels.setup

Climate Model Interface

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: Array{String,1} = Array{String,1}(undef, 0)
inputs :: Array{String,1} = Array{String,1}(undef, 0)
outputs :: Array{String,1} = Array{String,1}(undef, 0)
status :: Array{String,1} = Array{String,1}(undef, 0)
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and clean the run directory (via rm)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0,configuration="anonymous",options=Array{String,1}(undef, 0))
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.compile — Function

compile(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0)
setup(tmp)
compile(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
launch(tmp)

ClimateModels.monitor — Function

monitor(x)

Show x.status[end] by default.

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
monitor(tmp)

Git Support

ClimateModels.init_git_log — Function

init_git_log(x :: AbstractModelConfig)

Create log subfolder, initialize git, and commit initial README.md

ClimateModels.git_log_msg — Function

git_log_msg(x :: AbstractModelConfig,msg,commit_msg)

Add message msg to the log/README.md file and git commit.

ClimateModels.git_log_fil — Function

git_log_fil(x :: AbstractModelConfig,fil,commit_msg)

Commit changes to file log/fil with message commit_msg. If log/fil is unknown to git (i.e. commit errors out) then try adding log/fil first.

ClimateModels.git_log_prm — Function

git_log_prm(x :: AbstractModelConfig,msg,commit_msg)

Add files found in tracked_parameters/ (if any) to git log.

Cloud Support

ClimateModels.cmip — Function

cmip(institution_id,source_id,variable_id)

Access CMIP6 climate model archive (https://bit.ly/2WiWmoh) via AWS.jl and Zarr.jl and compute (1) time mean global map and (2) time evolving global mean.

This example was partly inspired by @rabernat 's https://bit.ly/2VRMgvl notebook

using ClimateModels
(mm,gm,meta)=cmip()
nm=meta["long_name"]*" in "*meta["units"]

using Plots
heatmap(mm["lon"], mm["lat"], transpose(mm["m"]),
        title=nm*" (time mean)")
plot(gm["t"][1:12:end],gm["y"][1:12:end],xlabel="time",ylabel=nm,
     title=meta["institution_id"]*" (global mean, month by month)")
display.([plot!(gm["t"][i:12:end],gm["y"][i:12:end], leg = false) for i in 2:12])