ClimateModels.jl

Note: package in early development stage; breaking changes remain likely.

This package provides a uniform interface to climate models of varying complexity and completeness. Models that range from low dimensional to whole Earth System models can be run and/or analyzed via this framework.

It also supports e.g. cloud computing workflows that start from previous model output available over the internet. Version control, using git, is included to allow for workflow documentation and reproducibility.

Example Workflows That Run Models

one stochastic path (0D)

a shallow water model (2D)

the MIT general circulation model (3D, Ocean, Atmosphere, etc)

Remote Access To Model Output

The initial example accesses CMIP6 model output from cloud storage, via AWS and Zarr, to compute temperature maps and time series.

Climate Model Interface

The climate model interface is based on a data structure (ModelConfig) and a series of methods (incl., setup and launch). The defaults assume that the model is a Julia package to be downloaded from a URL (setup), and run via Pkg.Test (launch). But the key point is that everything can be customized to e.g. use popular models previously written in Fortran or C.

Leveraging the interface in real world application essentially means :

Define a concrete type ModelConfig (optional).

Customize interface methods to best suit your chosen model.

At first, one can skip the type definition (#1 above) and may only want to customize setup and launch for #2 (see 0D and 2D examples).

But for routine use of e.g. a popular model it is suggested that the customized interface elements be incorporated in a dedicated package maintained inpdependently by developers / users familiar with that model.

This approach is illustrated in the general circulation model example in these docs, which uses the customized interface elements provided by MITgcmTools.jl for MITgcm .

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and clean the run directory (via rm)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0,configuration="anonymous")
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.compile — Function

compile(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0)
setup(tmp)
compile(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
launch(tmp)

ClimateModels.monitor — Function

monitor(x)

Show x.status[end] by default.

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
monitor(tmp)

API Reference

ClimateModels.ModelConfig

ClimateModels.build

ClimateModels.clean

ClimateModels.cmip

ClimateModels.compile

ClimateModels.git_log_fil

ClimateModels.git_log_init

ClimateModels.git_log_msg

ClimateModels.git_log_prm

ClimateModels.git_log_show

ClimateModels.launch

ClimateModels.monitor

ClimateModels.setup

Git Support

The setup method normally calls git_log_init to set up a temporary run folder with a git enabled subfolder called log. This allows for recording a workflow steps e.g. through the other functions listed here.

ClimateModels.git_log_init — Function

git_log_init(x :: AbstractModelConfig)

Create log subfolder, initialize git, and commit initial README.md

ClimateModels.git_log_msg — Function

git_log_msg(x :: AbstractModelConfig,msg,commit_msg)

Add message msg to the log/README.md file and git commit.

ClimateModels.git_log_fil — Function

git_log_fil(x :: AbstractModelConfig,fil,commit_msg)

Commit changes to file log/fil with message commit_msg. If log/fil is unknown to git (i.e. commit errors out) then try adding log/fil first.

ClimateModels.git_log_prm — Function

git_log_prm(x :: AbstractModelConfig,msg,commit_msg)

Add files found in tracked_parameters/ (if any) to git log.

ClimateModels.git_log_show — Function

git_log_show(x :: AbstractModelConfig)

Show git log.

Cloud Support

There are various ways that model output gets archived, distributed, and retrieved from the internet. In some cases downloading data can be the most convenient approach. In others it can be more advantageous to compute in the cloud and only download final results for plotting (e.g. cmip).

ClimateModels.cmip — Function

cmip(institution_id,source_id,variable_id)

Access CMIP6 climate model archive (https://bit.ly/2WiWmoh) via AWS.jl and Zarr.jl and compute (1) time mean global map and (2) time evolving global mean.

This example was partly inspired by @rabernat 's https://bit.ly/2VRMgvl notebook

using ClimateModels
(mm,gm,meta)=cmip()
nm=meta["long_name"]*" in "*meta["units"]

using Plots
heatmap(mm["lon"], mm["lat"], transpose(mm["m"]),
        title=nm*" (time mean)")
plot(gm["t"][1:12:end],gm["y"][1:12:end],xlabel="time",ylabel=nm,
     title=meta["institution_id"]*" (global mean, month by month)")
display.([plot!(gm["t"][i:12:end],gm["y"][i:12:end], leg = false) for i in 2:12])