User Manual

Here we document key functionalities offered in ClimateModels.jl

Climate Model Interface

Tracked Worklow Framework

Cloud + On-Premise File Support

Climate Model Interface

The climate model interface is based on the ModelConfig data structure and a series of methods like setup, build, and launch. The typical sequence is shown just below where f is a function that (1) receives a ModelConfig as its only input argument, and (2) gets called via launch.

julia> using ClimateModels
julia> f=ClimateModels.RandomWalker
julia> MC=ModelConfig(model=f)
julia> setup(MC)
julia> build(MC)
julia> launch(MC)

The ModelConfig called MC is summarized using the show method which here reveals that f is just an alias for ClimateModels.RandomWalker.

  ID            = 6e8444b3-220f-402e-9224-dbd11b513790
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/6e8444b3-220f-402e-9224-dbd11b513790
  log subfolder = /tmp/6e8444b3-220f-402e-9224-dbd11b513790/log
  task(s)       = RandomWalker

The run folder name can be accessed directly using pathof and its content inspected using readdir.

2-element Vector{String}:
 "RandomWalker.csv"
 "log"

Here launch(MC) has completed, RandomWalker.csv is a file that was generated by function f, and log is the log subfolder that was created by setup.

The workflow log is retrieved using the log function. log(MC) in this example would give:

5-element Vector{String}:
 "9947f79 initial setup"
 "63e7e77 add Project.toml to log"
 "fa2a84e add Manifest.toml to log"
 "89359f5 task started [64e58e36-303a-4903-8ecd-2270f6583f0b]"
 "63e8857 (HEAD -> main) task ended [64e58e36-303a-4903-8ecd-2270f6583f0b]"

This highlights that Project.toml and Manifest.toml for the environment being used have been archived. This happens during setup to document all dependencies and make the workflow reproducible.

Generalization

A key point is that everything can be customized to, e.g., use popular models previously written in Fortran or C just as simply. This typically involves defining a new concrete type of AbstractModelConfig and then providing customized build and/or setup methods as discussed below.

To start, let's distinguish amongst ModelConfigs on the basis of their model variable type :

normal user mode is when model is a String or a Function

package developer mode is when model is a Pkg.Types.PackageSpec

Normal User Mode

The simplest way to use the ClimateModels.jl interface is to specify model directly as a function, and use defaults for everything else. This approach may be best suited to pure Julia models or as a first step. It is illustrated in random walk. In CMIP6 the model name is provided as a String and the main Function is specified as the configuration instead.

Note

Once the initial launch call has completed, it is always possible to add workflow steps via put! and launch.

Often though, there are benefits to defining a custom setup and/or build method. One can then simply define a concrete type of AbstractModelConfig using ModelConfig as a blueprint. This is the recommended approach when another languange like Fortran, C++, or Python is involved. Thiis is illustrated with Hector, FaIR, SPEEDY, and MITgcm.

Defining a concrete type of AbstractModelConfig can also be practical with pure Julia model, e.g. to speed up launch, generate ensembles, facilitate checkpointing, etc. That's the case in the Oceananigans.jl example.

The idea in the longer term is that for popular models the customized interface elements would be provided via a dedicated package. They would thus be maintained independently by developers and users most familiar with each model. This approach is demonstrated in MITgcmTools.jl for MITgcm which provides its own suite of examples using the ClimateModels.jl interface.

Package Developer Mode

The defining feature of this approach is that the PackageSpec specification of model makes setup install the chosen package using Pkg.develop. This allows for developing a package or using an unregistered package in the context of ClimateModels.jl. There are two cases:

if configuration is left undefined then launch will run the package test suite using Pkg.test as in this example (code link, download link)

if configuration is provided as a Function then launch will call it as illustrated in the ShallowWaters.jl example (code link, download link)

Note

As an exercise, can you turn ShallowWaters.jl example into a normal user mode example?

Git / Log Support

The setup method normally calls log to create a temporary run folder with a git enabled subfolder called log. This allows for recording each workflow step, using log functions listed below.

Base.log — Function

log(x :: AbstractModelConfig)

Show the record of git commits that have taken place in the log folder.

log(x :: AbstractModelConfig, commit_msg :: String; 
             fil="", msg="", init=false, prm=false)

Keyword arguments work like this

init==true : create log subfolder, initialize git, and commit initial README.md

prm==true : add files found in input or tracked_parameters/ (if any) to git log

!isempty(fil) : commit changes to file log/$(fil) with message commit_msg. If log/$(fil) is unknown to git (i.e. commit errors out) then try adding log/$(fil) first.

and are mutually exclusive (i.e., use only one at a time).

using ClimateModels

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f,inputs=Dict("NS"=>100))

setup(tmp)
build(tmp)
launch(tmp)

msg="update tracked_parameters.toml (or skip if up to date)"
log(tmp,msg,fil="tracked_parameters.toml")
log(tmp)

Cloud / File Support

There are various ways that numerical model output gets archived, distributed, and retrieved from the internet. In some cases downloading data can be the most convenient approach. In others it can be more advantageous to compute in the cloud and only download final results for plotting.

ClimateModels.jl comes equiped with packages that read popular file formats used in climate modeling and science. Downloads.jl, CSV.jl, DataFrames.jl, NetCDF.jl, Zarr.jl, and TOML.jl are thus readily available when you install ClimateModels.jl. For instance, one can read the CSV file generated before as

julia> fil=joinpath(pathof(MC),"RandomWalker.csv")
julia> ClimateModels.CSV.File(fil) |> ClimateModels.DataFrame

or, maybe preferably, as

julia> fil=joinpath(pathof(MC),"RandomWalker.csv")
julia> CSV=ClimateModels.CSV
julia> DataFrame=ClimateModels.DataFrame
julia> CSV.File(fil) |> DataFrame

For additional examples covering other file formats, please refer to the IPCC report and CMIP6 archive notebooks and code links.

API Reference

ClimateModels.ModelConfig

Base.Filesystem.readdir

Base.log

Base.pathof

Base.show

ClimateModels.build

ClimateModels.clean

ClimateModels.cmip

ClimateModels.launch

ClimateModels.setup

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig)

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)
build(tmp)
launch(tmp)

Base.show — Function

show(io::IO, z::AbstractModelConfig)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
show(tmp)

Base.pathof — Function

pathof(x::AbstractModelConfig)

Returns the run directory path for x ; i.e. joinpath(x.folder,string(x.ID))

Base.Filesystem.readdir — Function

readdir(x::AbstractModelConfig)

Same as readdir(pathof(x)).

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and rm the run directory (pathof(x))

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

ClimateModels.cmip — Function

cmip(institution_id,source_id,variable_id)

Access CMIP6 climate model archive (https://bit.ly/2WiWmoh) via AWS.jl and Zarr.jl and compute (1) time mean global map and (2) time evolving global mean.

This example was partly inspired by @rabernat 's https://bit.ly/2VRMgvl notebook

using ClimateModels
(mm,gm,meta)=cmip()
nm=meta["long_name"]*" in "*meta["units"]

using Plots
heatmap(mm["lon"], mm["lat"], transpose(mm["m"]),
        title=nm*" (time mean)")
plot(gm["t"][1:12:end],gm["y"][1:12:end],xlabel="time",ylabel=nm,
     title=meta["institution_id"]*" (global mean, month by month)")
display.([plot!(gm["t"][i:12:end],gm["y"][i:12:end], leg = false) for i in 2:12])