User Manual

Here we document key functionalities offered in ClimateModels.jl

Climate Model Interface

Tracked Worklow Framework

Cloud + On-Premise File Support

Climate Model Interface

The climate model interface is based on the ModelConfig data structure and a series of methods like setup, build, and launch. The typical sequence is shown just below where f is a function that (1) receives a ModelConfig as its only input argument, and (2) gets called via launch.

using ClimateModels #hide
f=ClimateModels.RandomWalker #hide
MC=ModelConfig(model=f)
setup(MC)
build(MC)
launch(MC)

The ModelConfig called MC is summarized using the show method which here reveals that f is just an alias for ClimateModels.RandomWalker. The run folder name can be accessed directly using pathof.

  ID            = 6e8444b3-220f-402e-9224-dbd11b513790
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/6e8444b3-220f-402e-9224-dbd11b513790
  log subfolder = /tmp/6e8444b3-220f-402e-9224-dbd11b513790/log
  task(s)       = RandomWalker

The content of the run folder is readily inspected using readdir.

1-element Vector{String}:
 "log"

Here launch(MC) has completed, RandomWalker.csv is a file that was generated by function f, and log is the log subfolder that was created by setup.

The workflow log is retrieved using the log function.

log(MC) #hide

This highlights that Project.toml and Manifest.toml for the environment being used have been archived. This happens during setup to document all dependencies and make the workflow reproducible.

Generalization

A key point is that everything can be customized to, e.g., use popular models previously written in Fortran or C just as simply. This typically involves defining a new concrete type of AbstractModelConfig and providing a customized build method to compile the model.

This approach is illustrated in the examples as well as in the suite of general circulation model examples that uses the customized interface elements provided by MITgcmTools.jl for MITgcm.

The idea in the longer term is that for popular models the customized interface elements would be provided via a dedicated package (e.g. MITgcmTools.jl). These customized interfaces would thus be maintained independently by developers and users most familiar with each model.

Git / Log Support

The setup method normally calls log to set up a temporary run folder with a git enabled subfolder called log. This allows for recording each workflow step, using log functions listed below.

Base.log — Function

log(x :: AbstractModelConfig)

Show the record of git commits that have taken place in the log folder.

log(x :: AbstractModelConfig, commit_msg :: String; 
             fil="", msg="", init=false, prm=false)

init=true : create log subfolder, initialize git, and commit initial README.md

prm=true : add files found in input or tracked_parameters/ (if any) to git log

!isempty(fil) : commit changes to file log/fil with message commit_msg. If log/fil is unknown to git (i.e. commit errors out) then try adding log/fil first.

using ClimateModels

f=ClimateModels.RandomWalker
i=ClimateModels.OrderedDict(); i["NS"]=100

tmp=ModelConfig(model=f,inputs=i)
setup(tmp)
build(tmp)
launch(tmp)

log(tmp,
    "update tracked_parameters.toml (or skip)", 
    fil="tracked_parameters.toml")
ClimateModels.@suppress log(tmp)
isa(tmp,AbstractModelConfig)

Cloud / File Support

There are various ways that model output gets archived, distributed, and retrieved from the internet (e.g., see cmip). In some cases downloading data can be the most convenient approach. In others it can be more advantageous to compute in the cloud and only download final results for plotting.

ClimateModels.jl comes equiped with packages that read popular file formats used in climate modeling and science. Downloads.jl, CSV.jl, DataFrames.jl, NetCDF.jl, Zarr.jl, and TOML.jl are thus readily available when you install ClimateModels.jl.

For example, one can read the CSV file generated by ClimateModels.RandomWalker as follows:

fil=joinpath(pathof(MC),"RandomWalker.csv")
ClimateModels.CSV.File(fil) |> ClimateModels.DataFrame
show(ans) #hide

API Reference

ClimateModels.ModelConfig

Base.Filesystem.readdir

Base.log

Base.pathof

Base.show

ClimateModels.build

ClimateModels.clean

ClimateModels.cmip

ClimateModels.launch

ClimateModels.setup

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig)

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)
build(tmp)
launch(tmp)

Base.show — Function

show(io::IO, z::AbstractModelConfig)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
show(tmp)

Base.pathof — Function

pathof(x::AbstractModelConfig)

Returns the run directory path for x ; i.e. joinpath(x.folder,string(x.ID))

Base.Filesystem.readdir — Function

readdir(x::AbstractModelConfig)

Same as readdir(pathof(x)).

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and rm the run directory (pathof(x))

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

ClimateModels.cmip — Function

cmip(institution_id,source_id,variable_id)

Access CMIP6 climate model archive (https://bit.ly/2WiWmoh) via AWS.jl and Zarr.jl and compute (1) time mean global map and (2) time evolving global mean.

This example was partly inspired by @rabernat 's https://bit.ly/2VRMgvl notebook

using ClimateModels
(mm,gm,meta)=cmip()
nm=meta["long_name"]*" in "*meta["units"]

using Plots
heatmap(mm["lon"], mm["lat"], transpose(mm["m"]),
        title=nm*" (time mean)")
plot(gm["t"][1:12:end],gm["y"][1:12:end],xlabel="time",ylabel=nm,
     title=meta["institution_id"]*" (global mean, month by month)")
display.([plot!(gm["t"][i:12:end],gm["y"][i:12:end], leg = false) for i in 2:12])