Manual / User Guide

Here we document key functionalities offered in ClimateModels.jl

Climate Model Interface

The climate model interface is based on a data structure (ModelConfig) and a series of methods like setup, build, and launch. The default assumption is that the model is either 1) a Julia package to be downloaded from a URL within setup using Pkg.develop, and run within launch using Pkg.test or 2) a Julia function to be called with a ModelConfig argument.

A key point is that everything can be customized to e.g. 1) use a custom workflow instead of Pkg.test or 2) use popular models previously written in Fortran or C just as simply. The latter typically involves calling a build method to compile the model between setup and launch.

Leveraging the interface in real world application essentially means :

Define a concrete type ModelConfig (optional).

Customize interface methods to best suit your chosen model.

At first, one can skip the type definition (#1 above) and may only want to customize setup and launch for #2 (see examples 1, 2, and 3).

However, the idea is that for routine use of e.g. a popular model the customized interface elements would be provided via a dedicated package (e.g. MITgcmTools.jl). These customized interfaces would thus be maintained independently by developers and users most familiar with each model.

This approach is illustrated in the general circulation model example that uses the customized interface elements provided by MITgcmTools.jl for MITgcm.

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Suppressor, OrderedCollections
inputs=OrderedDict(); inputs["NS"]=1000;
tmp=ModelConfig(model=ClimateModels.RandomWalker,inputs=inputs)
setup(tmp)

build(tmp)
launch(tmp)
git_log_fil(tmp,"tracked_parameters.toml","update tracked_parameters.toml (or skip)")
@suppress git_log_show(tmp)
isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0,configuration="anonymous")
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.compile — Function

compile(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels, Pkg
tmp0=PackageSpec(url="https://github.com/JuliaOcean/AirSeaFluxes.jl")
tmp=ModelConfig(model=tmp0)
setup(tmp)
compile(tmp)

isa(tmp,AbstractModelConfig)

# output

true

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and clean the run directory (via rm)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
launch(tmp)

ClimateModels.monitor — Function

monitor(x)

Show x.status[end] by default.

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
monitor(tmp)

Git Support

The setup method normally calls git_log_init to set up a temporary run folder with a git enabled subfolder called log. This allows for recording each workflow step, using functions listed here for example.

ClimateModels.git_log_init — Function

git_log_init(x :: AbstractModelConfig)

Create log subfolder, initialize git, and commit initial README.md

ClimateModels.git_log_msg — Function

git_log_msg(x :: AbstractModelConfig,msg,commit_msg)

Add message msg to the log/README.md file and git commit.

ClimateModels.git_log_fil — Function

git_log_fil(x :: AbstractModelConfig,fil,commit_msg)

Commit changes to file log/fil with message commit_msg. If log/fil is unknown to git (i.e. commit errors out) then try adding log/fil first.

ClimateModels.git_log_prm — Function

git_log_prm(x :: AbstractModelConfig,msg,commit_msg)

Add files found in tracked_parameters/ (if any) to git log.

ClimateModels.git_log_show — Function

git_log_show(x :: AbstractModelConfig)

Show git log.

Cloud Support

There are various ways that model output gets archived, distributed, and retrieved from the internet. In some cases downloading data can be the most convenient approach. In others it can be more advantageous to compute in the cloud and only download final results for plotting (e.g. cmip).

ClimateModels.cmip — Function

cmip(institution_id,source_id,variable_id)

Access CMIP6 climate model archive (https://bit.ly/2WiWmoh) via AWS.jl and Zarr.jl and compute (1) time mean global map and (2) time evolving global mean.

This example was partly inspired by @rabernat 's https://bit.ly/2VRMgvl notebook

using ClimateModels
(mm,gm,meta)=cmip()
nm=meta["long_name"]*" in "*meta["units"]

using Plots
heatmap(mm["lon"], mm["lat"], transpose(mm["m"]),
        title=nm*" (time mean)")
plot(gm["t"][1:12:end],gm["y"][1:12:end],xlabel="time",ylabel=nm,
     title=meta["institution_id"]*" (global mean, month by month)")
display.([plot!(gm["t"][i:12:end],gm["y"][i:12:end], leg = false) for i in 2:12])

API Reference

ClimateModels.ModelConfig

ClimateModels.build

ClimateModels.clean

ClimateModels.cmip

ClimateModels.compile

ClimateModels.git_log_fil

ClimateModels.git_log_init

ClimateModels.git_log_msg

ClimateModels.git_log_prm

ClimateModels.git_log_show

ClimateModels.launch

ClimateModels.monitor

ClimateModels.setup