User Manual

Here we document key functionalities offered in ClimateModels.jl

Climate Model Interface

Tracked Worklow Framework

Cloud + On-Premise File Support

Climate Model Interface

The interface ties the ModelConfig data structure with methods like setup, build, and launch.

f=ClimateModels.RandomWalker
MC=ModelConfig(model=f)

  ID            = 74541fe0-b1ea-413c-b040-7551ac95572b
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/74541fe0-b1ea-413c-b040-7551ac95572b
  log subfolder = /tmp/74541fe0-b1ea-413c-b040-7551ac95572b/log

The typical sequence is shown below. Here f is a function that receives a ModelConfig as its only input argument. It gets called via launch and generates a file called RandomWalker.csv.

setup(MC)
build(MC)
launch(MC)

For convenience, run executes all three steps at once. Using the simplified ModelConfig constructor, we then just write:

run(ModelConfig(f))

  ID            = 375d34dd-ac46-49f4-a16a-c6459a6a929f
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/375d34dd-ac46-49f4-a16a-c6459a6a929f
  log subfolder = /tmp/375d34dd-ac46-49f4-a16a-c6459a6a929f/log

Note

Once the initial model run has completed, it is always possible to add workflow steps via put! and launch.

The run folder name and its content can be viewed using pathof and readdir, respectively.

pathof(MC)

"/tmp/74541fe0-b1ea-413c-b040-7551ac95572b"

readdir(MC)

2-element Vector{String}:
 "RandomWalker.csv"
 "log"

The show method for ModelConfig is demonstrated below.

show(MC)

  ID            = 74541fe0-b1ea-413c-b040-7551ac95572b
  model         = RandomWalker
  configuration = anonymous
  run folder    = /tmp/74541fe0-b1ea-413c-b040-7551ac95572b
  log subfolder = /tmp/74541fe0-b1ea-413c-b040-7551ac95572b/log

The log subfolder was created earlier by setup. The log function retrieves the workflow log.

log(MC)

5-element Vector{String}:
 "63b0202 initial setup"
 "079bd3a add Project.toml to log"
 "ff4f530 add Manifest.toml to log"
 "1a2c369 task started [85dbd099-aa49-4adb-82d4-1f9c73e8e1bb]"
 "95e857b (HEAD -> main) task ended   [85dbd099-aa49-4adb-82d4-1f9c73e8e1bb]"

This highlights that Project.toml and Manifest.toml for the environment being used have been archived. This happens during setup to document all dependencies and make the workflow reproducible.

Generalization

A key point is that everything can be customized to, e.g., use popular models previously written in Fortran or C just as simply.

This typically involves defining a new concrete type of AbstractModelConfig and then providing customized setup, build, and/or launch methods.

To start, let's distinguish amongst ModelConfigs on the basis of their model variable type :

normal user mode is when model is a String or a Function.

package developer mode is when model is a PackageSpec (see below).

Normal User Mode

The simplest way to use the ClimateModels.jl interface is to specify model directly as a function, and use defaults for everything else, as illustrated in random walk. Alternatively, the model name can be provided as a String and the main Function as the configuration, as in CMIP6.

Often, however, one may want to define custom setup, build, or launch methods. To this end, one can define a concrete type of AbstractModelConfig using ModelConfig as a blueprint. This is the recommended approach when other languanges like Fortran or Python are involved (e.g., Hector, FaIR, SPEEDY, MITgcm).

Note

Defining a concrete type of AbstractModelConfig can also be practical with pure Julia model, e.g. to speed up launch, generate ensembles, facilitate checkpointing, etc. That's the case in the Oceananigans.jl example.

For popular models the customized interface elements can be provided via a dedicated package. This may allow them to be maintained independently by developers and users most familiar with each model. MITgcmTools.jl does this for MITgcm. It provides its own suite of examples that use the ClimateModels.jl interface.

Package Developer Mode

The defining feature of this approach is that the PackageSpec specification of model makes setup install the chosen package using Pkg.develop. This allows for developing a package or using an unregistered package in the context of ClimateModels.jl. There are two cases:

if configuration is left undefined then launch will run the package test suite using Pkg.test as in this example (code link)

if configuration is provided as a Function then launch will call it as illustrated in the ShallowWaters.jl example (code link)

Note

As an exercise, can you turn ShallowWaters.jl example into a normal user mode example?

Tracked Worklow Support

The setup method normally calls log to create a temporary run folder with a git enabled subfolder called log. This allows for recording each workflow step, using log functions listed below.

Cloud And File Support

There are various ways that numerical model output gets archived, distributed, and retrieved from the internet. In some cases downloading data can be the most convenient approach. In others it can be more advantageous to compute in the cloud and only download final results for plotting.

ClimateModels.jl comes equiped with packages that read popular file formats used in climate modeling and science. Downloads.jl, CSV.jl, DataFrames.jl, NetCDF.jl, Zarr.jl, and TOML.jl are thus readily available when you install ClimateModels.jl. For instance, one can read the CSV file generated before as

fil=joinpath(pathof(MC),"RandomWalker.csv")
CSV.File(fil) |> DataFrame

"100×2 DataFrame"

For additional examples covering other file formats, please refer to the IPCC report and CMIP6 archive notebooks and code links.

API Reference

ClimateModels.ModelConfig — Type

struct ModelConfig <: AbstractModelConfig

model :: Union{Function,String,Pkg.Types.PackageSpec} = "anonymous"
configuration :: Union{Function,String} = "anonymous"
options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
channel :: Channel{Any} = Channel{Any}(10) 
folder :: String = tempdir()
ID :: UUID = UUIDs.uuid4()

ClimateModels.ModelConfig — Method

ModelConfig(func::Function,inputs::NamedTuple)

Simplified constructor for case when model is a Function.

ClimateModels.PkgDevConfig — Function

PkgDevConfig(url::String,func::Function,inputs::NamedTuple)

Simplified constructor for case when model is a url (PackageSpec).

ClimateModels.setup — Function

setup(x)

Defaults to default_ClimateModelSetup(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)

ClimateModels.build — Function

build(x)

Defaults to default_ClimateModelBuild(x). Can be expected to be specialized for most concrete types of AbstractModelConfig

using ClimateModels
tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
build(tmp)

isa(tmp,AbstractModelConfig) # hide

ClimateModels.launch — Function

launch(x)

Defaults to default_ClimateModelLaunch(x) which consists in take!(x) for AbstractModelConfig. Can be expected to be specialized for most concrete types of AbstractModelConfig

f=ClimateModels.RandomWalker
tmp=ModelConfig(model=f)
setup(tmp)
build(tmp)
launch(tmp)

Base.run — Function

run(x :: AbstractModelConfig)

Shorthand for setup |> build |> launch, which returns AbstractModelConfig as output.

Utilities

Base.show — Function

show(io::IO, z::AbstractModelConfig)

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
show(tmp)

Base.log — Function

log(x :: AbstractModelConfig)

Show the record of git commits that have taken place in the log folder.

log(x :: AbstractModelConfig, commit_msg :: String; 
             fil="", msg="", init=false, prm=false)

Keyword arguments work like this

init==true : create log subfolder, initialize git, and commit initial README.md

prm==true : add files found in input or tracked_parameters/ (if any) to git log

!isempty(fil) : commit changes to file log/$(fil) with message commit_msg. If log/$(fil) is unknown to git (i.e. commit errors out) then try adding log/$(fil) first.

and are mutually exclusive (i.e., use only one at a time).

MC=run(ModelConfig(ClimateModels.RandomWalker,(NS=100,)))
MC.inputs[:NS]=200
msg="update tracked_parameters.toml (or skip if up to date)"
log(MC,msg,fil="tracked_parameters.toml",prm=true)
log(MC)

Base.pathof — Function

pathof(x::AbstractModelConfig)

Returns the run directory path for x ; i.e. joinpath(x.folder,string(x.ID))

Base.Filesystem.readdir — Function

readdir(x::AbstractModelConfig)

Same as readdir(pathof(x)).

ClimateModels.clean — Function

clean(x :: AbstractModelConfig)

Cancel any remaining task (x.channel) and rm the run directory (pathof(x))

tmp=ModelConfig(model=ClimateModels.RandomWalker)
setup(tmp)
clean(tmp)

Various

ClimateModels.cmip — Function

cmip(institution_id,source_id,variable_id)

Access CMIP6 climate model archive via AWS.jl and Zarr.jl and compute (1) time mean global map and (2) time evolving global mean.

This example was partly inspired by this notebook.

using ClimateModels
(mm,gm,meta)=cmip()
nm=meta["long_name"]*" in "*meta["units"]

using Plots
heatmap(mm["lon"], mm["lat"], transpose(mm["m"]),
        title=nm*" (time mean)")
plot(gm["t"][1:12:end],gm["y"][1:12:end],xlabel="time",ylabel=nm,
     title=meta["institution_id"]*" (global mean, month by month)")
display.([plot!(gm["t"][i:12:end],gm["y"][i:12:end], leg = false) for i in 2:12])

ClimateModels.notebooks.open — Function

pluto_url="http://localhost:1234/"
#plut_url="https://notebooks.gesis.org/binder/jupyter/user/juliaclimate-notebooks-vx8glon1/pluto/"
#pluto_url="https://ade.ops.maap-project.org/serverpmohyfxe-ws-jupyter/server-3100/pluto/"
nbs=notebooks.list()

path=joinpath(tempdir(),"nbs")
notebooks.download(path,nbs)

ii=1
notebook_url=nbs.url[ii]
#notebook_path=joinpath("tutorials","jl",nbs.folder[ii],nbs.file[ii])
notebook_path=joinpath(path,nbs.folder[ii],nbs.file[ii])

notebooks.open(pluto_url,notebook_url=notebook_url)
#notebooks.open(pluto_url,notebook_path)